refactor: :art: remove others entities types on file and save bkp to reference
